+++ Gruntfile.js
@@ -1,21 +1,84 @@
 "use strict";
 
 module.exports = function(grunt) {
-    
-    require("load-grunt-tasks")(grunt);
-    
-    require('time-grunt')(grunt);
-     
+  require("load-grunt-tasks")(grunt);
 
-    require('load-grunt-config')(grunt, {
-        jitGrunt: true
-    });
-	
-	grunt.loadNpmTasks('grunt-contrib-jade');
-    grunt.loadNpmTasks('grunt-contrib-jshint');
-    grunt.loadNpmTasks('grunt-browser-sync');
-    
-    
-   // grunt.registerTask("default", ["browserSync", "watch"]);
-	
+  grunt.initConfig({
+    sass: {
+      style: {
+        files: {
+          "css/style.css": "sass/style.scss"
+        }
+      }
+    },
+
+    postcss: {
+      options: {
+        processors: [
+          require("autoprefixer")({browsers: [
+            "last 1 version",
+            "last 2 Chrome versions",
+            "last 2 Firefox versions",
+            "last 2 Opera versions",
+            "last 2 Edge versions"
+          ]})
+        ]
+      },
+      style: {
+        src: "css/*.css"
+      }
+    },
+
+    browserSync: {
+      server: {
+        bsFiles: {
+          src: [
+            "*.html",
+            "css/*.css"
+          ]
+        },
+        options: {
+          server: ".",
+          watchTask: true,
+          notify: false,
+          open: true,
+          ui: false
+        }
+      }
+    },
+      
+    jade: {
+      
+        options: {
+          processName: function(filename) {
+            return filename.toUpperCase();
+          },
+          client: false,
+          runtime: true,
+          pretty: true,
+          compileDebug: false,
+          extension: false
+        },
+        all: {
+            files: [{
+                expand: true,
+                cwd: 'jade/',
+                src: '**/*.jade',
+                dest: '',
+                ext: '.html'
+            }]
+        }
+        
+    },
+
+    watch: {
+      files: ["sass/**/*.{scss,sass}","jade/**/*.{jade}"],
+      tasks: ["sass", "postcss","jade"],        
+      options: {
+        spawn: false
+      }
+    }
+  });
+
+  grunt.registerTask("serve", ["browserSync", "watch"]);
 };
