--- grunt/aliases.yaml
@@ -1,30 +1,0 @@
-default:
-  description: 'Default (production) build'
-  tasks:
-    - prod
-dev:
-  description: 'Development build'
-  tasks:
-    - 'concurrent:devFirst'
-    - 'concurrent:devSecond'
-img:
-  description: 'Image tasks'
-  tasks:
-    - 'concurrent:imgFirst'
-devimg:
-  description: 'Development build and image tasks'
-  tasks:
-    - dev
-    - img
-prod:
-  description: 'Production build'
-  tasks:
-    - 'concurrent:prodFirst'
-    - 'concurrent:prodSecond'
-    - img
-pack:
-  description: 'Jade tasks'
-  tasks:
-    - jade
-    - sass:dev
-    - autoprefixer
