--- grunt/browsersync.js
@@ -1,19 +1,0 @@
-module.exports = {
-    
-browserSync: {
-            dev: {
-                bsFiles: {
-                    src : [
-                        'css/*.css',
-                        '*.html'
-                    ]
-                },
-                options: {
-                    watchTask: true,
-                    server: '.'
-                }
-            }
-        }
-    
-    
-};
