--- grunt/concurrent.js
@@ -1,32 +1,0 @@
-module.exports = {
-
-    // Опции
-    options: {
-        limit: 3
-    },
-
-    // Задачи разработки
-    devFirst: [
-        'clean',
-        'jshint'
-    ],
-    devSecond: [
-        'sass:dev',
-        'uglify'
-    ],
-
-    // Производственные задачи
-    prodFirst: [
-        'clean',
-        'jshint'
-    ],
-    prodSecond: [
-        'sass:prod',
-        'uglify'
-    ],
-
-    // Задачи изображений
-    imgFirst: [
-        'imagemin'
-    ]
-};
