--- grunt/imagemin.js
@@ -1,10 +1,0 @@
- module.exports = {
-    all: {
-        files: [{
-            expand: true,
-            cwd: '',
-            src: ['img/*.{png,jpg,gif}'],
-            dest: 'dist/'
-        }]
-    }
-};
