--- grunt/jshint.js
@@ -1,10 +1,0 @@
-module.exports = {
-
-    options: {
-        reporter: require('jshint-stylish')
-    },
-
-    main: [
-        'js/*.js'
-    ]
-};
