--- grunt/postcss.js
@@ -1,21 +1,0 @@
-module.exports =  {
-    // Development settings
-
-    postcss: {
-      options: {
-        processors: [
-          require("autoprefixer")({browsers: [
-            "last 1 version",
-            "last 2 Chrome versions",
-            "last 2 Firefox versions",
-            "last 2 Opera versions",
-            "last 2 Edge versions"
-          ]})
-        ]
-      },
-      style: {
-        src: "css/*.css"
-      }
-    }
-    
-};
