--- grunt/uglify.js
@@ -1,11 +1,0 @@
-module.exports = {
-    all: {
-        files: [{
-            expand: true,
-            cwd: 'js/',
-            src: '**/*.js',
-            dest: 'dist/js',
-            ext: '.min.js'
-        }]
-    }
-};
