+++ index.html
@@ -2,32 +2,31 @@
 <html>
   <head>   
     <meta charset="utf-8"/>
-    <title>PINK adaptive    </title>
+    <meta name="viewport" content="width=device-width,initial-scale=1"/>
+    <title>PINK adaptive </title>
     <link href="css/style.css" rel="stylesheet"/>
   </head>
   <body>     
     <header class="main-header"> 
       <div class="main-header__top">       
-        <nav class="main-nav">          
-          <div class="main-nav__logo"> <a href="#"><img src="img/img_logo.png" width="75" height="22" alt="PINK"/></a></div>
-          <div class="main-nav__menu"><a href="#"><i class="main-nav__icon">   </i></a>
-            <ul class="main-nav__items">
-              <li>Информация</li>
-              <li>Возможности </li>
-              <li>Отзывы</li>
-              <li>Стоимость</li>
-            </ul>
-          </div>
-        </nav>
-        <div class="intro">
-          <div class="intro__layout">
+        <div class="main-header__layout">
+          <nav class="main-nav">          
+            <div class="main-nav__logo"> <a href="#"><img src="img/img_logo.png" width="75" height="22" alt="PINK"/></a></div>
+            <div class="main-nav__menu"><a href="#"><i class="main-nav__icon">   </i></a>
+              <ul class="main-nav__items">
+                <li class="main-nav__item"> <a href="#" class="main-nav__link">Информация</a></li>
+                <li class="main-nav__item"> <a href="#" class="main-nav__link">Возможности</a></li>
+                <li class="main-nav__item"> <a href="#" class="main-nav__link">Отзывы</a></li>
+                <li class="main-nav__item"> <a href="#" class="main-nav__link">Стоимость</a></li>
+              </ul>
+            </div>
+          </nav>
+          <div class="intro">
             <h1 class="intro__header">Взгляните на жизнь иначе!                          </h1>
-          </div>
-        </div>
-        <div class="download">
-          <div class="download__layout"><a href="#" class="btn btn--upload">Скачать приложение</a>
-            <div class="download__items"><a href="#" class="download__item"><i class="download__icon download__icon--apple"></i></a><a href="#" class="download__item">             <i class="download__icon download__icon--android"></i></a><a href="#" class="download__item">             <i class="download__icon download__icon--microsoft"></i></a></div>
-            <p class="download__description">Доступно для iPhone, iPad, Android, Windows Phone, OS X, Windows 8</p>
+            <div class="download">         <a href="#" class="btn btn--upload">Скачать приложение</a>
+              <div class="download__items"><a href="#" class="download__item"><i class="download__icon download__icon--apple"></i></a><a href="#" class="download__item">             <i class="download__icon download__icon--android"></i></a><a href="#" class="download__item">             <i class="download__icon download__icon--microsoft"></i></a></div>
+              <p class="download__description">Доступно для iPhone, iPad, Android, Windows Phone, OS X, Windows 8</p>
+            </div>
           </div>
         </div>
       </div>
