+++ jade/photo.jade
@@ -2,6 +2,7 @@
 html
   head    
     meta(charset="utf-8")
+    meta(name="viewport" content="width=device-width,initial-scale=1")    
     title PINK adaptive    
     link(href="css/style.css" rel="stylesheet")
   body
