+++ package.json
@@ -15,30 +15,20 @@
   "devDependencies": {
     "autoprefixer": "6.3.3",
     "editorconfig-tools": "0.1.1",
-    "grunt": "^0.4.5",
-    "grunt-browser-sync": "^2.2.0",
-    "grunt-concurrent": "^2.2.1",
-    "grunt-contrib-clean": "^1.0.0",
-    "grunt-contrib-imagemin": "^1.0.0",
+    "grunt": "0.4.5",
+    "grunt-browser-sync": "2.2.0",
     "grunt-contrib-jade": "^1.0.0",
-    "grunt-contrib-jshint": "^1.0.0",
-    "grunt-contrib-uglify": "^1.0.0",
     "grunt-contrib-watch": "0.6.1",
     "grunt-postcss": "0.7.2",
-    "grunt-sass": "^1.1.0",
-    "jshint-stylish": "^2.1.0",
-    "load-grunt-config": "^0.19.1",
+    "grunt-sass": "1.1.0",
     "load-grunt-tasks": "3.4.1"
   },
   "scripts": {
     "test": "editorconfig-tools check *",
-    "build": "grunt sass && grunt postcss",
+    "build": "grunt sass && grunt postcss && grunt jade",
     "start": "npm run build && grunt serve"
   },
   "engines": {
     "node": "^5.0.0"
-  },
-  "dependencies": {
-    "time-grunt": "^1.3.0"
   }
 }
