+++ sass/blocks/download.scss
@@ -1,3 +1,15 @@
 .download{
     
+  width: 300px;
+    
+  height: 270px;
+  outline: 1px solid rgba(0, 150, 150, 1);
+    
+  &:before{
+
+    content: "download";
+    background-color: rgba(0, 150, 150, .4);
+       
+  }
+    
 }
