+++ sass/blocks/main-nav.scss
@@ -1,3 +1,64 @@
 .main-nav{
     
+  outline: 1px solid yellow;
+
+  &:before{
+    
+    content: "main-nav";
+    background-color: rgba(150, 150, 50, .3);
+
+  }
+    
+  &:after{
+
+      @include clearfix;
+
+  }
+    
+  &__layout{
+   
+    @include layout();
+        
+  }
+
+  &__logo{
+
+    width: 75px;
+    height: 22px;
+    float: left;
+    outline: 1px solid yellow;
+    padding: 22px 0;
+
+  &:before{        
+
+    content: "__logo";
+    background-color: rgba(180, 180, 50, .3);
+
+   }
+
+ }
+
+  &__menu{
+
+    float: right;
+    height: 22px;    
+    width: 75px;
+    padding: 21px 0; 
+    outline: 1px solid yellow;  
+      
+    &:before{        
+
+      content: "__menu";
+      background-color: rgba(180, 180, 50, .3);
+
+    }      
+
+  }
+    
+  &__items{
+
+    display: none;
+
+  }
+    
 }
