+++ sass/mixin.scss
@@ -1,16 +1,30 @@
 /* Миксины */
 
 @mixin layout{
-    
+      
   padding: 0 20px;    
   margin: 0 auto;
+    
+  @media(min-width: 700px) and (max-width: 1199px) {
+  
+    padding: 0 40px;
+      
+  }     
+    
+  @media(min-width: 1200px){
+  
+    padding: 0 130px;
+      
+  }      
     
 }
 
 @mixin contain-width{
     
   margin: 0 auto; 
-  width: 320px;
+
+  width: 100%;
+    
     
 }
 
@@ -19,5 +33,13 @@
   border-top: 40px solid $color-white;
   border-right:320px solid $color-gray;
   margin-bottom: 53px;
+    
+}
+
+@mixin clearfix{
+    
+    display: table;
+    clear: both;
+    content: '';    
     
 }
