+++ sass/style.scss
@@ -61,8 +61,23 @@
 
     
 
-body{
+body{    
     
-  width: $width-body;    
-      
+    @media(max-width: 699px){
+        
+        min-width: 320px;
+        
+    }
+    @media(min-width: 700px) and (max-width: 1199px){
+        
+        min-width: 700px;
+        
+    }
+    @media(min-width: 1200px){
+        
+        max-width: 1200px;
+        
+    }
+  
+  width: $width-body;      
 }
