+++ sass/variables.scss
@@ -1,5 +1,9 @@
 $width-body: 100%;
 
+$width-content-mobile: 320px;
+$width-content-tablet: 700px;
+$width-content-desktop: 1200px;
+
 /* шрифты */
 
 $open-sans-light: 'Open Sans Light', 'Arial', open-sans;
